requirements.txt
numpy==2.0.1
boto3==1.34.150
botocore==1.34.150
fastapi==0.110.3
passlib==1.7.4
pydantic==1.10.4
python-dotenv==1.0.0
python_jose==3.3.0
python-multipart==0.0.5
uvicorn==0.20.0
email-validator==1.3.1
requests==2.28.2
pandas==2.2.2
bcrypt==4.0.1
openpyxl==3.1.5
uuid==1.30

